✨ puzzle screen 코드 구조 분리 및 점수 계산 로직 수정
